APP/requirements.txt
amqp==2.6.0
aniso8601==7.0.0
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.5
Django==3.0.7
django-environ==0.4.5
django-graphql-jwt==0.3.1
django-graphql-playground==1.0
future==0.18.2
graphene==2.1.8
graphene-django==2.10.1
graphql-core==2.3.2
graphql-relay==2.0.1
importlib-metadata==1.6.1
kombu==4.6.10
promise==2.3
PyJWT==1.7.1
pytz==2020.1
redis==3.5.3
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.3.1
vine==1.3.0
zipp==3.1.0
